requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
distlib==0.3.6
Django==4.2.3
filelock==3.12.2
idna==3.4
mysqlclient==2.2.0
pipenv==2023.7.4
platformdirs==3.8.1
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
whitenoise==6.5.0
